requirements.txt
Pillow==12.0.0
pandas==2.3.3
